Use userfaultfd's SIGBUS feature for concurrent compaction
With the threading-based implementation the mutator threads and
userfaultfd worker threads have to be alternatively scheduled when
a missing page is accessed by the former. OTOH, with SIGBUS feature
the mutator gets a SIGBUS signal on accessing a missing page.

For response time the latter is expected to be significantly better than
the former. With a microbenchmark on host SIGBUS feature is up to 10x
better than the alternative.

Bug: 160737021
Test: art/test/testrunner/testrunner.py --host
Change-Id: I6f8d05690e23b70f9517e9e1929af3006b9960bb