Remove OatWriter::DexFileSource
The AddDexFileSource methods open the input file, read the header,
and do minimal amount verification.  However, this is what
DexFileLoader does (with the full dex verification disabled).
Therefore we can remove the code and use DexFileLoader directly.

Likewise, the AddDexFileSource manually handled the lifespan
of the various types of open input files, but the DexFileLoader
already can do automatically via the DexFileContainer.
Therefore we can just replace DexFileSource with DexFile.

Test: ./art/test.py -b --host --all-cdex_level --64
Change-Id: I72185f53d79785a1df0dcf8dee10ab88e606894d