Remove `kMultiFPRegistersWidened` as it is false for all architectures.
In `BuildNativeCallFrameStateMachine`, `kMultiFPRegistersWidened`
constant was responsible for the cases when 32-bit floats must be
"widened" (converted to double before they are written to a 64-bit
register). This is not true for any of the supported architectures, so
there's no need to maintain that logic (previously it was true for
mips32, obsoleted by commit 41b605c5ad4b06ea127ac56c6e3a4c92e8913efd
that removed mips support).

Test: lunch aosp_riscv64-userdebug \
  && art/tools/buildbot-build.sh --target
Change-Id: Idd8ecf6d5383909d6b94f6b8bf3093b201b6b699