Set the RTI to a default value when we don't have enough information
In some cases, the bytecode doesn't provide enough information for
us to set a valid RTI. In those cases set the default value of
Object. For BoundType instructions, we can instead use its
UpperBound (as long as it is valid).

gtest update: We need the handles_ to be non-null for
ReferenceTypePropagation so create them and keep them alive until
we finish creating the CFG.

Bug: 252799892
Fixes: 252799892
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the apps in the bug
Change-Id: If2b6a85dc74032bd95bd8032ed5874b4bdb94b5c