Refactor DexFileLoader
The first parameter of the various Open methods specifies the data source,
that we intend to load the dex file(s) from.  This creates large number
of overloads when multiplied by diversity of the other arguments as well.

Move the date source parameters to the constructor of DexFileLoader.
Specifically, the constructor just creates the right DexFileContainer,
and the rest of the loader is independent of the data source used.

This removes large amount of the overloads and well as large amount
of copy-pasted code. Majority of ArtDexFileLoader has been removed.

One notable side effect is that zip files are open as one single
container now, which is shared between it's DexFiles.

Bug: 266950186
Test: ./art/test.py -b --host --optimizing --64
Change-Id: I6580b49e65441eec93a7e0124be23bd8c859904a