Don't reset the_trace_ before flushing per-thread buffers
When we want to stop tracing, we set the_trace_ to nullptr and then
maintain a copy of a local trace to compute trace summary and to flush
the trace to the file. This works fine before we used per-thread
buffers. Since we started using per-thread buffer, this can cause a
problem if a thread gets destroyed before we clear the per-therad buffer
at a later step. When a therad gets destroyed with a valid buffer we try
to flush it and expect the_trace_ to be valid too. This CL avoids this
by resetting per-therad buffer before clearing the_trace_. We do this in
SuspendAll scope to avoid any races.

Bug: 259258187
Test: art/test.py
Change-Id: I8f2e59bad1508e26ee43302ee1d619a35bb2f660