Revert "Regenerate the ART test files (2023-02-10)."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0c2a848-f124-416d-bd41-7f89347c736a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5791995bdd1cbc06bd84ad9b274639dbce5454ca