Don't break domination when code sinking
We can create an example where the try always throws
and its catch flows to the beginning of the try,
effectively having an infinite loop. Our code sinking
algorithm assumes we don't have an infinte loop and
generates a bad graph.

To fix this, we can choose not to perform the
heuristic that tries to choose a `target_block` outside
of loops when we have a try catch in the graph.

Bug: 269927803
Fixes: 269927803
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the app in the bug
Change-Id: Id563a2e87798d74833081e254b2856937bdf7e2c