Don't try to hoist an instruction out of a loop it belongs in
We can create an example where the try always throws
and its catch flows to the beginning of the try,
effectively having an infinite loop. Our code sinking
algorithm assumes we don't have an infinte loop and
generates a bad graph.

To fix this, we can improve our code sinking heuristics to
not try to hoist an instruction out of a loop it was
already in.

Bug: 269927803
Fixes: 269927803
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the app in the bug
Change-Id: Id563a2e87798d74833081e254b2856937bdf7e2c