riscv64: further deduplicate `art_quick(_static)_invoke_stub` code.
The two stubs (static and non-static) are different only in that the
non-static stub loads `this` argument in the A1 register. This CL
factors out the jump table with LOADREG* routines and moves it to the
common part of the stub implementation.

Also, use an extra parameter to the macro so that we can keep human
readable label names rather than numeric labels.

Test: run ART test 001-HelloWorld on a Linux RISC-V VM:
  lunch aosp_riscv64-userdebug

  export ART_TEST_SSH_USER=ubuntu
  export ART_TEST_SSH_HOST=localhost
  export ART_TEST_SSH_PORT=10001
  export ART_TEST_ON_VM=true

  . art/tools/buildbot-utils.sh
  art/tools/buildbot-build.sh --target

  # Create, boot and configure the VM.
  art/tools/buildbot-vm.sh create
  art/tools/buildbot-vm.sh boot
  art/tools/buildbot-vm.sh setup-ssh  # password: 'ubuntu'

  art/tools/buildbot-cleanup-device.sh
  art/tools/buildbot-setup-device.sh
  art/tools/buildbot-sync.sh

  art/test.py --target -r --no-prebuild --ndebug --no-image \
    --64 --interpreter 001-HelloWorld

Bug: b/271573990
Change-Id: I2cd5d61192167419e5cf40671a22ac1461037aaf