Refactor DataBegin() and DataEnd() calculation in DexFile.
In this context, the data range is defined at the memory region
that most of the dex offsets are relative to. It is same as the
whole dex file for standard dex, but special for compact dex.

Push the calculation of this memory range into DexFile internals,
since it will different for multidex in the future (where it will
be the memory range of the whole mulitdex container).

Bug: 266950186
Test: test.py -b --host --all-cdex_level --64
Change-Id: I1e4ad9af975fd15c24c1d89a5fc51c63ea562091