Revert "Generate runtime app images into the app cache directory."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3446f243-fccc-4dc1-ac34-841b93df62ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21c00e1cbc884d03e982fe61e84385629838724a