Revert "Fix run-gtests.sh for device."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e41e374-4506-4b23-baa1-6a158395002e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie44a28943db902bb47a045d354ed31aa430f090b