[Sim] Enable custom linker relocations for Arm64 EPs
Add custom relocations for symbols accessed by Arm64 entrypoints to
allow those symbols to be accessed when the Arm64 ELF object file is
linked against a host (x86_64) libart.so. This allows Arm64 EP code
to exist in the x86_64 library and later be run using the ART
simulator.

Instead of letting the compiler create relocations to symbols on
instructions (e.g: "bl foo") we manually create a placeholder
relocation (R_AARCH64_NONE) to symbols on a separate literal
(.dword 0x0000000000000000) in the instruction stream. This literal
is the target location for the relocation and, once linked, will
contain the PC relative address of the symbol (e.g: "foo") which is
loaded into a register and used.

The placeholder relocation is expected to be replaced before link
time with the appropriate PC relative x86_64 relocation type:
R_X86_64_PC64.

To use these entrypoints also initialise them for the simulator
instead of the x86_64 entrypoints.

Test: test.py --host --target

Change-Id: I413895e5abc762d4f0c681f9dcfea10f8c8c49c7