Fix the inconsistency of Thread.setPriority()
When calling Thread.setPriority(), in some cases, or in
the kernel, such as our custom emulator, the native thread
priority setting will fail.

However, the priority of the Java thread precedes the native
thread priority setting. This causes the native thread priority
setting to fail after the Java thread priority setting succeeds.
In this scenario, you need to update the Java thread priority
to the correct value.

Test: art/test/051-thread

Change-Id: I867224429529bca9f31ecd787fb6cd0a5658419c
Signed-off-by: wangzijie <wangzijie@eswincomputing.com>