Revert "Add isRecord() and getRecordComponents() API to Class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e6608f5-bde1-4a04-b374-64fecb82d4a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I684a233b436b9ead6413af68a66eaac97836512d