Revert "Add isRecord() and getRecordComponents() API to Class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87155d14-fcb5-4cdf-b4cb-d7bac22f0b43). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I606360f169a706bb6b90e50a5d6e2294ee2c04dc