Revert "Add isRecord() and getRecordComponents() API to Class"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a836b09f-85d6-4d18-a2c3-6e601cb45dea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie8f9a0abf2136f6dc7831362e5232c87bac5c7e6