Make art modules apex_available to test_broken_com.android.art
libart-broken needs to compile against the source of libdexfile, not
stubs. The stubs/source logic is complicated, but one of the ways it
can be affected is that if both modules have the same apex_available
values, they will be allowed to depend on the source of each other.
libart-broken needs to be apex_available to test_broken_com.android.art,
so test_broken_com.android.art needs to be added to libdexfile as well.
This change then propagates virally though all the art modules.

This is needed now because we're making a change to allow building the
non-apex variants of apex dependencies. The apex variant of
libart-broken did not have this issue because the stubs/source rules
are different for apex variants.

Bug: 254205429
Test: cd art && mm with aosp/2462194
Test: art/libnativebridge/tests/runtests.sh --skip-target
Change-Id: Ia3517b70808d64ed2b33d68e2280b514bdfef6b6