Revert "Revert^4 "Add the mainline framework boot image extensio..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8524f342-b206-4df5-87dd-a90054543c45). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I682b3c424e189798f4bbbb8722a5a178e9160834