Revert "Rewrite `LocalReferenceTable`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02af463e-4bc0-41b0-80db-7beec5ef7ea0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I34d9682c0c1374ee5c1183abe83f115ea8ef85ae