Move native asm out of quick_entrypoints*.S
There are some entrypoints (both from quick_entrypoints_*.S and
jni_entrypoints_*.S) that are called from a native context (in native
calling convention and stack frame layout, as opposed to the quick
equivalent) and will always return to a native context. Therefore
move these entrypoints to a new native_entrypoints_*.S file to better
identify them as staying in a native context.

This is also a prerequisite to simulating all quick entrypoints, when
the native code ISA (e.g: x86_64) is not the same as the quick code
ISA (e.g: Arm64), as it allows us to compile and link quick
entrypoints differently to native entrypoints.

Test: test.py --host --target
Change-Id: Ia798e675b95b93ee867fba491dfae3ae6b3d1983