Re-add ComposHostTestCases
We want to run them as a presubmit here, since ART changes can cause
the tests to fail and we want to catch that.

Commit c3cb031bb1c5a8cda83bfab601560cf94c29f4aa removed them from
presubmit since they were flaky on cuttlefish, but they are now
disabled on cuttlefish.

Now run them using avf-presubmit, which will ensure they run on some
real devices.

Bug: 264496291
Test: Run regen-test-files, see expected output
Change-Id: I9ff8120d2694305210fe7d9416ce346310ba722a