riscv64: Move shadow stack register from X18 to X3(GP).
This is following a corresponding change in LLVM and bionic. Note that
we don't use X18 yet to make the migration path easier. It can be used
in the future after all compiler prebuilts are updated.

This commit also drops art thread register TR from `kSaveEverything`
frame (which allowas us to keep frame size unchanged). Dropping TR
should have been done in https://r.android.com/2430733, and all the
comments suggested that it was dropped, so this commit brings comments
in sync with the actual code.

Bug: 277909695
Bug: b/271573990
Test: run ART run-tests on a Linux RISC-V VM, observe that they pass all
      except one expected failure 051-thread.

  lunch aosp_riscv64-userdebug

  export ART_TEST_SSH_USER=ubuntu
  export ART_TEST_SSH_HOST=localhost
  export ART_TEST_SSH_PORT=10001
  export ART_TEST_ON_VM=true

  . art/tools/buildbot-utils.sh
  art/tools/buildbot-build.sh --target

  # Create, boot and configure the VM.
  art/tools/buildbot-vm.sh create
  art/tools/buildbot-vm.sh boot
  art/tools/buildbot-vm.sh setup-ssh  # password: 'ubuntu'

  art/tools/buildbot-cleanup-device.sh
  art/tools/buildbot-setup-device.sh
  art/tools/buildbot-sync.sh

  art/test.py --target -r --no-prebuild --ndebug --no-image \
    --64 --interpreter 984-obsolete-invoke

Change-Id: I52ee3c8e82ac734638ba9f1546620d26b1c20c73