Remove art_debug_defaults custom soong module type
Replace it with a soong config variable-based module type.

The other art module types will probably be kept and converted to
bazel macros, but the defaults ones are tricker because we run the
defaults mutator before bp2build is run.

Bp2build can convert the art module types to bazel, but it sees the
cc modules that the art modules make after the art modules have affected
their properties. So they essentially "bake in" the art module changes
into the generated BUILD file. This causes the generated BUILD files
to be different depending on which product was lunched. We don't want
those differences because it means we can't check in the generated
BUILD files.

Bug: 247785938
Test: Verified ninja files don't change before/after this cl. (they actually do, but only in comments) Also tested a build with ART_DEBUG_OPT_FLAG=-O0
Change-Id: I7a28eb6455f05b3b846586270b32c6b8e1be200c