Revert "Don't generate an image for debuggable apps."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5747d09f-e6f7-4bd0-85e3-e0c9e1e7121d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7331be19866bae768b8af91c9fa5515b8b152f3e