Revert "Add some debugging output for b/261719949."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/231508bf-2e50-406a-91b7-8a1a28024b59). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic36c8b863964bb1671e8895800a166f7fee933b2