Throw ClassFormatError for invalid @Record annotation
ART has 2 ways to deal with invalid record class. For exmaple,
non-final j.l.Record subclass is not treated as a record class,
i.e. Class.isRecord() returns false.

The second way is to throw ClassFormatError if the dex format
doesn't follow the spec. This CL implements this part.

Record class is a new feature in Android U.

Bug: 272698028
Test: CtsLibcoreOjTestCases:test.java.io.Serializable.records
Test: CtsLibcoreTestCases:crossvmtest
Test: CtsLibcoreTestCases:libcore.java.lang.ClassTest
Test: CtsLibcoreTestCases:libcore.java.lang.RecordTest
Test: CtsLibcoreOjTestCases:test.java.lang.reflect.records
Change-Id: I3eacc9b4fcef47596704e7f1c386368041db0bf0