Revert "Don't enable intrinsic optimizations in debuggable runtime"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80542304-50b2-4aa4-8d1d-995222ef4ad6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib44b38d84b47c504f6d89fa7147c9cea6e40f408