Revert "Don't enable intrinsic optimizations in debuggable runtime"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64009992-8645-492a-93eb-e037a59714f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7fbc71e27962a02328c3004e8c87e7e3be6ffdcc