Revert "Don't enable intrinsic optimizations in debuggable runtime"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f8c8bfc5-2297-4219-8f42-ff6e4db5657e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie792d07a7715dae12191bd01bef84f5898139e07