Revert "Remove the codegen property"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/88fe93a5-fc12-490e-8914-e81bb746b001). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62231cc8635bd50ce1ba9a94e41a090acce5dcdd