Revert "Remove the codegen property"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3440cc9b-39a8-433b-99ad-532494f62c1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3ac06ac7e7823897da950d44bdf2927517134a33