Revert "Add more debugging logs for b/279117765."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed946d00-b45d-4e29-bdbc-e1097f12daee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd8365a4a8cabc3c81bbbbbedf407df683449016