NativeAllocation tweaks
Define kCheckImmediatelyThreshold in terms of kNotifyNativeInterval.
We previously raised kNotifyNativeInterval without adjusting it,
making it possible to go for much longer periods with checking
whether we should GC. This effectively reduces
kCheckImmediatelyThreshold, by about a factor of 2 in production,
and more on host. This is unlikely to fix the failures.

Have 004-NativeAllocations produce much more detailed output on
failure.

Bug: 123921229
Test: TreeHugger
Change-Id: Ibe61ea16be8a5f25ec241f682ef3b4958a50116f