Add Arm64 CPU variants with SVE support
Add tested CPU variants with SVE support, and set the default SVE
vector length to 128-bit, which is validated when compiling for an
SVE-supporting CPU variant on the target.

We do this validation in Arm64InstructionSetFeatures::FromHwcap
which is called during JitCompiler initialization and parsing dex2oat
options in target builds when we pass --instruction-set-variant option
that is always true for production scenarios.

Note that this CL adds instruction-set feature support for SVE, but ART
code generation for SVE also depends on the kArm64AllowSVE compiler
constant.

Target tests on arm64 with sve enabled, when passing
"cortex-<x2|x3|a710|a715|a510>" to --instruction-set-variant, or when
passing "sve" to --instruction-set-features.

Test: testrunner.py --target --optimizing --jit

Authors: Richard Neill <richard.neill@arm.com>
         Konstantin Baladurin <konstantin.baladurin@arm.com>

Change-Id: I1950d0d9fe862e07341855c1b5159da6a7373a41