Revert "Add pageinfo tool for memory page analysis"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5ca2a7a-0256-4f38-ac83-29ea32f7af32). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id021cc1f83a033a260a42eff2fe4e67ee071ff0d