Revert "Static link gtests with ART libraries"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21827566-47f8-43e7-8809-bda2eaf05f26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3e0b9dfdbf6e898988123a9cbc8558f508ad8e12