Revert "Static link gtests with ART libraries"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8ba21f4-d734-4a0b-847d-c67d5f080f6f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83adcb0407ee944dd4a964271721ea6345a5b8e1