Revert "Static link gtests with ART libraries"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df54e30f-f6e1-4f83-a982-a7222474b028). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36a9edcdc074c8a1b9d4314f9f26ea5c5e1a0bf9