Revert "Static link gtests with ART libraries"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6b7f4fd-1b50-41ac-b85c-a913352d5469). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I96efd4ac73d99c4a32edf27f76644672cc2784e2