Add check for read barriers in Checker
One of the consequences of having the new uffd GC is that the read
barrier is completely removed, causing some Checker tests (e.g.,
1004-checker-volatile-ref-load) that are explicitly checking for
its present to fail. This CL emits the information on what type of
read barrier is present in the Control-Flow Graph and adds a
readBarrierType('...') check that allows to have different assertions
for different read barrier types.

Bug: 283780888
Test: atest -a art-run-test-1004-checker-volatile-ref-load
Test: art/tools/checker/run_unit_tests.py
Change-Id: Ie57839f933b82c5134697368db9143b7b23af599