Revert "Reland "Use memset/madv_free instead of dontneed in fore..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a9e82fc-51b1-4591-a40b-0cbaa05d5a53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I632b4a0902e90be9e7a87790f5ef72a859109ce3