Use libfstab to determine whether to put artifacts in dalvik-cache.
Before this change, we used PackageState#isSystem (a.k.a.
ApplicationInfo#FLAG_SYSTEM) to determine this. Unfortunately, some OEMs
don't set this flag for some apps because the flag is used for making
security decisions in some places, although the document says it
shouldn't be.

After this change, we use libfstab as a general solution.

Bug: 287958783
Test: atest art_standalone_artd_tests
Test: atest ArtServiceTests
Change-Id: I0ef596f7009c3d19f458d58c922041ab8a7fbf5f
Merged-In: I0ef596f7009c3d19f458d58c922041ab8a7fbf5f