Revert "Reland "Use memset/madv_free instead of dontneed in fore..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c6b7724-4599-447d-a152-555ee2fde0e0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I27bbc05b979f68b684300f5f3e0402c15b8e76e6