Change the meaning of --interpreter option in testing scripts.
Previously --interpreter used to run the tests on the slow C++
switch-interpreter (the analogue of -Xint option to dalvikvm). Now it
runs the tests on the fast assembly interpreter, Nterp (the analogue of
-Xusefit:false option to dalvikvm).

On the other hand, --interp-ac now means what --interpreter previously
meant (running on the slow switch interpreter).

Bug: 271573990

Test: `art/test.py --host -r 001-HelloWorld` with the following
    modifications that ensure ART runs in the intended mode:
      - no modifications: 10/10 tests pass
      - deliberaly break Nterp: 2/10 tests pass (only --interp-ac ones)
      - deliberately break JIT: 8/10 tests pass (only --jit tests fail)
Change-Id: If2ec1df96b65c7e2dd7e81dff3ee0378bde62751