Revert "Reduce multidex checksum to single scalar value."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe8e3eba-8d83-4a32-9506-aae0550c0731). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd40898e606ce33022f2595d01f1e49813cb71f1