Revert "Replace libc++.so.1 with libc++.so"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0734c747-6232-435c-b5d0-be1d9b0b6abb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e157595d85e219d8e1feb5ca63890e914f0a500