Unify handling of thread infos in method tracing
Information about threads is recorded differently in streaming and
non-streaming modes. In streaming we just record the information when we
see the first event from the thread. In non-streaming we record all the
threads by recording them when any thread is getting destroyed and also
all active threads when stopping tracing. This is prone to races and
requires few workarounds to prevent them. So just handle them similar to
streaming case.

Bug: 259258187
Test: art/test.py --stream
Change-Id: Ib8efdb9fb875c1f3a4acfd3e9c3879e02a63847d