Pin kLargeObjectAlignment
This patch is preparing for making kPageSize non-constexpr
in a future patch.

The value used to be initialized with page size value.

Now, as the page size becomes variable, this should be changed, as
kLargeObjectAlignment is a template parameter. Change it to fixed value
of maximum page size which should work for all page sizes.

With page size agnosticism mode switched off the maximum supported
page size would be 4K, so in that mode kLargeObjectAlignment would
have same value as without this change.

Test: Same as for I5430741a8494b340ed7fd2d8692c41a59ad9c530.
      The whole patches chain was tested as a whole.
Change-Id: I138f6351c5420217b9be8bb839cf68a48223af8d