[WIP] Pin kLargeObjectAlignment
This patch is preparing for making kPageSize non-constexpr
in a future patch.

The value used to be dependent on page size.
Now, as the page size becomes variable, it should be changed, as
kLargeObjectAlignment is a template parameter. Change it to fixed value
of 16K therefore requiring 16K alignment for 4K configuration too.

Consider setting to 4k.

Test: Same as for Id1c18004346ba5c6c94e02cdf8b0b0bb3b99af70.
      The whole patches chain was tested as a whole.
Change-Id: I138f6351c5420217b9be8bb839cf68a48223af8d