[WIP] Use CHECK_ALIGNED_PARAM for variable page size
This patch is part of the chain preparing for making kPageSize
non-constexpr in a future patch.

Since kPageSize is going to be a non-constexpr value, it can't be a
template parameter anymore. Consequently CHECK_ALIGNED(..., kPageSize),
expressions implemented via a template function IsAligned<>, have to be
replaced with CHECK_ALIGNED_PARAM(..., kPageSize).

Same is done for DCHECK_ALIGNED.

Test: Same as for Id1c18004346ba5c6c94e02cdf8b0b0bb3b99af70.
      The whole patches chain was tested as a whole.
Change-Id: I0e634dab62f2b44d078d7199d5b5feab945077b9