Make kPageSize-derived values non-constexpr
This patch is part of the chain preparing for making kPageSize
non-constexpr in a future patch.

Since kPageSize is going to be a non-constexpr value, it can't be
used to initialize constexpr values anymore. Remove constexpr specifier
and replace it with const type qualifier for the derived values.

However, for configuration with page size agnosticism switched off,
effects on performance are expected to be negligible as the values
are still effectively constant so compiler should be able to easily
deduce that and propagate the constants.

Test: Same as for I5430741a8494b340ed7fd2d8692c41a59ad9c530.
      The whole patches chain was tested as a whole.
Change-Id: I901d6b0504355ed290c76569c4472aed7414956c