[Sim] Build target boot image for simulator
The simulator requires the target boot image to run and correctly
simulate target tests however the normal ART target build no longer
builds the target boot image on the host. Instead it is built on the
target device which the simulator cannot use.

To resolve this, add a build step to build the target boot image on
the host if the simulator is being built.

Test: test_art_simulator.sh -t
Change-Id: I8c9f2a1c2e3f14bbc48bc3f2eef18a682ce5557b