Flush the trace entries on a thread pool worker
Flushing the trace entries from the per-thread buffer takes considerable
amount time which causes the process to be less responsive when
flushing. So move the flushing to a dedicated worker pool thread so that
the actual thread can continue. This is done only when flushing entries
in streaming mode. In non-streaming mode, we need to flush entries from
all the threads to see if we have space in the common buffer for any new
entries. So, we flush all the entries synchronously to keep the
implementation simple.

Bug: 259258187
Test: art/test.py --stream --trace / art/test.py --trace
Change-Id: I4912c0249a40747fd6cd3863c8617f0bab335267