[WIP] Make kMinLargeObjectThreshold a fixed constant
This patch is part of the chain preparing for making kPageSize
non-constexpr in a future patch.

Before, kMinLargeObjectThreshold was defined as 3 page sizes.
Make it independent of page size - for now choosing the value
that would have been used in 4K configuration.

Choosing the value needs further investigation as it might not be
optimal.

This is required because the constant is inlined in an assembly
instruction.

Test: Same as for Id1c18004346ba5c6c94e02cdf8b0b0bb3b99af70.
      The whole patches chain was tested as a whole.
Change-Id: I8eb8d6de0ab59e63308e8ec8c3a88573c236e5f6