Revert "Add DEX SHA1 to oat files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a07d493-8ad6-48f3-bce4-fbd045abc2d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I004e67836e5a2a12bcff832897759144fb6c19ac