Revert "Only set ArtMethod as shared memory if we compile with a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c6654211-bc20-452e-a2c0-48826104f756). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3e2c0c74d2495aca98abac5b764ee6f40a73a55f