Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/877b2dbd-44fd-4660-8bd3-a091481429c9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6ad04ac5b1b068f14b63620923ad11be3b2c7f6