Allow libdexfile support to be used by platform
libdebuggerd depends on libunwind, which depends on libdexfile.
Since libdexfile is not available for platform, the build system transitively marked libdebuggerd as "not available for platform",
even though "libdexfile" dependency is excluded in ramdisk variant
of libunwind.

By marking libdexfile as platform available, the ramdisk copy of
libdebuggerd is now available for platform usage.


Test: th
Change-Id: I1cdd5bb0fb9338fb36fb8ec618f9b872ead44e9f