Revert "Only set ArtMethod as shared memory if we compile with a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b230e6ea-7715-4e27-90f6-3679668cb9b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib99330ab4e04db68fe9a8741fbed481686d284ff