Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c877a7e0-5a49-46a1-8477-9374f9070678). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I183f9d39c62bfd34582484b014d022e56e7c1744