Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0faccb56-a162-4488-b1d0-5e4f8eaabead). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b4e8ca9f537f4aa249d0bc9389d5d5f51b9cc23