Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/38bda831-5a8d-4a8d-962a-18ab50de75c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89bc1155afba36a64fea059b27d9226a2e1f8d55