Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74cd2a6b-0b27-47d4-bf6e-11e4a6ffcf66). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id22e84b8abde9a07e339a8a6c876b39985b38f17