Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b4ba3150-fb08-44bb-837a-5f71a06fe905). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I135a8d6bbe1a3deb5bbf4c54a60567c6bb79270b