Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f998c4db-c329-49e3-9a5c-85df6a0f76e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5bbf5ba74cc2b5db3e956b24b02034dfdfb9e173