Revert "Revert "Merge libart and libart-compiler into a single l..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1112a8f-b51b-4a51-8366-dd60d0bc324e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8826dfada35ebec018bafa7dcc95ceb97642c01c