Revert "Merge libart and libart-compiler into a single library"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fbbbc72-acdb-4c9f-99cc-42bd52a6f809). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae147fe89dbb1999aa575f91f30e1d2c7d9b71c6