DO NOT SUBMIT: Copy of aosp/2702853 for presubmit UX study.
Allow to compile and verify with a full CLC

In previous versions, we were expected to compile with a partial
class loader context, but now we are expected to have the full
CLC at compile time. Update ART to allow the new way of doing
things, while keeping it backwards compatible.

Test: m test-art-host-gtest-art_runtime_tests64
Test: Manually compiled in both the new way, and the backwards
      compatible way and made sure we are successfully loading
      the oat files
Bug: 293298276
Change-Id: I8dc538eef169e28d2810959849195661a90c2a52