Revert "riscv64: Disassemble floating point operations."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/235576c9-63fb-4344-890c-b6479eba1a5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6296f9f79007fd2da7aeadc352f15b93b3739c3