Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4be3081-e1b0-4aa0-bac4-50d5fb345826). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b38d59456a36dfe1a98465cac5d37737d019f15