Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/77eb0237-92e2-442a-b64c-bbe6311b7d21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I271485d7c9e356d44acb3d8e9f5d6d2c5522e86f