Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71a1b346-4168-4ef2-b8a0-1e0ef1a518a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a942b2dfbe3c2a1a25f687874f2076777a785c1