Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0628f7dd-7b89-48a8-9a4c-7ac0440fe1ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe12c82833149f70e2fc6014f7ceaae90a71e947