Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81f92ece-f21c-4edc-8ceb-90d9abb5dd1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia45408088589c993c3af1a731457b4e68d4de695