Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8eb49be-b697-4790-8446-ea1bbcfad033). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I037e7a2252f3e98a6ccd5c3601e7338f71139b5a