Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8097fa21-c9a4-4bb0-8e03-2a5a5bd1b6ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a3fa8b2b8801957f66d6d57a54d97ed4c54c82e