Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb447821-9d88-49f3-8b1b-b1c93a78284a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7da74012da237ca595680ad385abb20cf32cfec4