Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/922141ef-aa3e-4c8d-b7f6-0a2da1608a99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d3333d82ad42a020a2e603ea450bc84540b03b8