Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cbb573b8-9c1e-43f6-8144-55ed31955842). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36aa4d3c027244269227dc7f4a559bacfc0272f7