Revert "Revert^2 "Merge libart and libart-compiler into a single..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1198b6b5-921b-4941-88eb-a062e4db7c55). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I16dba24fd944a187dd8bcd082e6fca99d944946b